Day83_PortfolioWebsite/requirements.txt
Bootstrap_Flask==2.2.0
Flask==2.3.2
Flask_CKEditor==0.4.6
Flask_Login==0.6.2
Flask-Gravatar==0.5.0
flask_sqlalchemy==3.0.5
Flask_WTF==1.2.1
WTForms==3.0.1
Werkzeug==2.3.3
SQLAlchemy==2.0.19
python-dotenv==1.0.0
email_validator==2.1.0